Improves ScheduleList (defending against concurrent updates, faster search)
Currently, all use of ScheduleList is on the main thread. So, locking is
removed; defending against concurrent updates to warn early if that
expectation ever changes. If it does, improve ScheduleList to deal with it.

Search for position to insert/move items to is now via binary search, not
linear.